HTML Sanitizing OK
For now there's now way to sanitize automatically for the post model
because we can't set the getter / setter directly. This happens because
currently we are using the Database First method for generating model
data.

For now ONLY access the Post.content colum through the custom
SanitizedContent field.

Might need to think about a way to move to code first or mixed mode so
the Post model's getter and setter can be changed, resulting in safer
code.